+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Date format: (YYYY-MM-DD)
 * Fixed: The 'active shopkeepers' would not get properly cleaned up in some occasions (even on shopkeeper removal) if the shopkeeper mob got deleted or the shop object was no longer considered 'active' for some other reason. A side effect of this was that the shopkeeper entity would get respawned, even though it was not supposed to get spawned or even after the shopkeeper was already deleted.
 * Fixed: The DerivedSettings use the default value for the name-regex setting during initialization now to properly catch user errors during the subsequent setup after the config has already been loaded.
 * Fixed: The selling and book shops attempted to convert currency items into high currency items even if the high currency got disabled.
+* Fixed: Trading via shift-clicking while the player is charging a trident would allow the player to duplicate the trident.
 
 Internal changes:  
 * Added AbstractShopkeeper#tick which gets invoked roughly once per second for all shopkeepers in currently active chunks.